+++ projects/docs/stories/lists/fullList/models/options.tsx
@@ -30,7 +30,8 @@ export const fullListFormOptions: IFormBuilderProps<IFullListFormItem> = {
             columnsPicks: {
                 columns: ["awesome"],
                 pickType: FormListColumnsPickType.Without
-            }
+            },
+            resizableColumns: true
         },
         editorConfig: {
             title: (item, creation) => creation ? "New item" : `Edit '${item?.firstname}'`,
@@ -49,42 +50,6 @@ export const fullListFormOptions: IFormBuilderProps<IFullListFormItem> = {
                     { title: "Edit", iconName: "Edit", selectionMode: IFormBuilderListMenuItemSelectionMode.Single, action: () => editItem() },
                     { title: "Delete", iconName: "Delete", selectionMode: IFormBuilderListMenuItemSelectionMode.Single | IFormBuilderListMenuItemSelectionMode.Multi, action: () => deleteItems(deleteItemsPreProcess) },
                     { title: "Click here!", iconName: "NumberField", selectionMode: IFormBuilderListMenuItemSelectionMode.Multi, action: items => alert(items.length) },
-                    {
-                        title: 'Additional actions',
-                        iconName: 'SetAction',
-                        selectionMode: IFormBuilderListMenuItemSelectionMode.Single | IFormBuilderListMenuItemSelectionMode.Multi,
-                        action: () => {},
-                        subMenuItems: [
-                            {
-                                title: 'Set active',
-                                iconName: 'SetAction',
-                                selectionMode: IFormBuilderListMenuItemSelectionMode.Single | IFormBuilderListMenuItemSelectionMode.Multi,
-                                action: () => {},
-                                subMenuItems: [
-                                    {
-                                        title: 'Activate',
-                                        selectionMode: IFormBuilderListMenuItemSelectionMode.Single | IFormBuilderListMenuItemSelectionMode.Multi,
-                                        action: () => {}
-                                    },
-                                    {
-                                        title: 'Inactivate',
-                                        selectionMode: IFormBuilderListMenuItemSelectionMode.Single | IFormBuilderListMenuItemSelectionMode.Multi,
-                                        action: () => {}
-                                    },
-                                ],
-                            },
-                            {
-                                title: 'Move all requests',
-                                selectionMode: IFormBuilderListMenuItemSelectionMode.Single | IFormBuilderListMenuItemSelectionMode.Multi,
-                                action: () => {}
-                            },
-                            {
-                                title: 'Move all allocations',
-                                selectionMode: IFormBuilderListMenuItemSelectionMode.Single | IFormBuilderListMenuItemSelectionMode.Multi,
-                                action: () => {}
-                            },
-                        ],
-                    },
                 ]
             ),
         }